backend/push-notifications/requirements.txt
psycopg2-binary
PyJWT
pywebpush
